The sampler and the accounting follow: luminal::search_support
Three copies of the genome sampler (reference, CUDA-lite, test runtime)
were byte-identical, and the reporting plumbing beside it — progress
lines, refusal counts, stage timings, the early-stop predicate — was
byte-identical in the two that carry a loop. None of it chooses an
implementation, so none of it is runtime-specific.

Moved to `src/search_support.rs`: the sampler block (`ProducerIndex`,
`choose_position`, `sample_genome{,_reporting,_with_seed}`,
`flip_closes_cycle`, `mutate_genome{,_reporting,_with_seed}`) and its
`bufferize_cycle_tripwire`, which is the sampler's own invariant check
and what `sampler_tests` half exists to pin; `RefusalBreakdown` and
`SearchTimings` with their summaries; `SearchProgress`,
`CaptureAwareStderr`, `log_channel_enabled`, `parse_log_flag`,
`display_nanos`; `early_stop_exceeded`, which had a copy in each runtime
and a copy of its test in each too. `RefusalBreakdown` takes CUDA-lite's
form, the superset — the reference runtime gains a `timed_out` field it
leaves at zero, and a `, timed out 0` tail on a diagnostic string no
test reads.

Deliberately NOT moved, because they decide: `CompileOptions`,
`SearchOutcome`, the evaluators, the GA loop, the bucketed drivers,
`select_bucket`, `harness_search_options`, and CUDA-lite's finalists and
lattice. The loop has to price a plan in the middle of every iteration,
and pricing is the runtime — the reference one runs the plan, CUDA-lite
reads a heuristic or profiles a device. Sharing it means core calling
back out, which is the seam the no-trait ruling closed. Both loops stay
where they are.

Each runtime's `search` module re-exports what it used to define, so
every public path still resolves: `crate::search::early_stop_exceeded`
for CUDA-lite's profiler, `test_runtime::sampler::*` for the SCC sampler
board. `test_runtime::sampler` is now the core module under its old name,
like `extractor`.

No RNG-consuming code changed — the sampler moved verbatim and both
loops call it in the same order — and the seeded pins are unchanged:
cublaslt_election, cublaslt_bias_premise, cublaslt_contracts_cpu,
finalists_lattice, dim_buckets.
